+++ deployments/abi/GitcoinAttester.json
@@ -52,5 +52,32 @@
     "function upgradeTo(address newImplementation)",
     "function upgradeToAndCall(address newImplementation, bytes data) payable",
     "function verifiers(address) view returns (bool)"
+  ],
+  "0xe704": [
+    "event AdminChanged(address previousAdmin, address newAdmin)",
+    "event BeaconUpgraded(address indexed beacon)",
+    "event Initialized(uint8 version)",
+    "event OwnershipTransferred(address indexed previousOwner, address indexed newOwner)",
+    "event Paused(address account)",
+    "event Unpaused(address account)",
+    "event Upgraded(address indexed implementation)",
+    "event VerifierAdded(address verifier)",
+    "event VerifierRemoved(address verifier)",
+    "function addVerifier(address _verifier)",
+    "function initialize()",
+    "function owner() view returns (address)",
+    "function pause()",
+    "function paused() view returns (bool)",
+    "function proxiableUUID() view returns (bytes32)",
+    "function removeVerifier(address _verifier)",
+    "function renounceOwnership()",
+    "function revokeAttestations(tuple(bytes32 schema, tuple(bytes32 uid, uint256 value)[] data)[] multiRevocationRequest) payable",
+    "function setEASAddress(address _easContractAddress)",
+    "function submitAttestations(tuple(bytes32 schema, tuple(address recipient, uint64 expirationTime, bool revocable, bytes32 refUID, bytes data, uint256 value)[] data)[] multiAttestationRequest) payable returns (bytes32[])",
+    "function transferOwnership(address newOwner)",
+    "function unpause()",
+    "function upgradeTo(address newImplementation)",
+    "function upgradeToAndCall(address newImplementation, bytes data) payable",
+    "function verifiers(address) view returns (bool)"
   ]
 }
