+++ hardhat.config.ts
@@ -16,6 +16,17 @@ let config: HardhatUserConfig = {
         url: process.env.PROVIDER_URL as string,
       },
     },
+    "linea-goerli": {
+      chainId: 59140,
+      gasPrice: 114868572,
+      url: `https://linea-goerli.infura.io/v3/${process.env.INFURA_KEY ?? ""}`,
+      accounts: [process.env.DEPLOYER_PRIVATE_KEY ?? ""],
+    },
+    linea: {
+      chainId: 59144,
+      url: `https://linea-mainnet.infura.io/v3/${process.env.INFURA_KEY ?? ""}`,
+      accounts: [process.env.DEPLOYER_PRIVATE_KEY ?? ""],
+    },
   },
   etherscan: {
     apiKey: process.env.ETHERSCAN_API_KEY as string,
