+++ package.json
@@ -10,7 +10,7 @@
     "@types/mocha": ">=9.1.0",
     "chai": "^4.3.7",
     "dotenv": "^16.0.3",
-    "hardhat": "^2.13.0",
+    "hardhat": "^2.17.2",
     "hardhat-gas-reporter": "^1.0.8",
     "prettier-plugin-solidity": "^1.1.3",
     "solhint": "^3.4.1",
@@ -25,6 +25,8 @@
     "@openzeppelin/contracts-upgradeable": "^4.9.2",
     "@openzeppelin/upgrades-core": "^1.27.3",
     "@typechain/ethers-v6": "^0.4.0",
-    "ethers": "^6.6.3"
+    "ethers": "^6.6.3",
+    "linea-attestation-registry": "https://github.com/Consensys/linea-attestation-registry.git#v2",
+    "openzeppelin-contracts-upgradeable": "https://github.com/OpenZeppelin/openzeppelin-contracts-upgradeable.git#f34a3a7"
   }
 }
